refactor: replace `GlobalComponentStore` with `GlobalRouterComponentStore` implementation
